Split into separate files and added tsne mapping
